requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
motor>=3.1.2
pydantic>=1.10.7
email-validator>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.6.0
python-dotenv>=1.0.0
pymongo>=4.3.3
bson>=0.5.10
uuid>=1.30